moved existing sensor hardware test
